+++ jkube-kit/common/src/test/java/org/eclipse/jkube/kit/common/util/KubernetesHelperTest.java
@@ -20,7 +20,6 @@ import io.fabric8.kubernetes.api.model.EnvVar;
 import io.fabric8.kubernetes.api.model.EnvVarBuilder;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.Namespace;
-import io.fabric8.kubernetes.api.model.NamespaceBuilder;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
 import io.fabric8.kubernetes.api.model.Quantity;
 import io.fabric8.kubernetes.api.model.ReplicationControllerBuilder;
@@ -58,6 +57,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -414,18 +414,29 @@ public class KubernetesHelperTest {
     }
 
     @Test
-    public void testGetNamespaceFromKubernetesList() {
+    public void testGetNamespaceFromKubernetesListReturnsValidNamespace() {
         // Given
-        List<HasMetadata> entities = new ArrayList<>();
-        entities.add(new NamespaceBuilder().withNewMetadata().withName("ns1").endMetadata().build());
-        entities.add(new DeploymentBuilder().withNewMetadata().withName("d1").endMetadata().build());
+        Properties properties = new Properties();
+        properties.put("jkube.namespace", "ns1");
 
         // When
-        String namespace = KubernetesHelper.getNamespaceFromKubernetesList(entities);
+        String namespace = KubernetesHelper.getConfiguredNamespace(properties, "default");
 
         // Then
         assertNotNull(namespace);
         assertEquals("ns1", namespace);
+    }
+
+    @Test
+    public void testGetNamespaceFromKubernetesListReturnsNull() {
+        // Given
+        Properties properties = new Properties();
+
+        // When
+        String namespace = KubernetesHelper.getConfiguredNamespace(properties, "default");
+
+        // Then
+        assertEquals("default", namespace);
     }
 
     @Test