+++ jkube-kit/config/service/src/main/java/org/eclipse/jkube/kit/config/service/ApplyService.java
@@ -61,7 +61,6 @@ import org.eclipse.jkube.kit.common.util.KubernetesHelper;
 import org.eclipse.jkube.kit.common.util.OpenshiftHelper;
 import org.eclipse.jkube.kit.common.util.ResourceUtil;
 import org.eclipse.jkube.kit.common.util.UserConfigurationCompare;
-import org.eclipse.jkube.kit.config.access.ClusterAccess;
 import org.eclipse.jkube.kit.config.resource.JKubeAnnotations;
 import org.eclipse.jkube.kit.config.service.kubernetes.KubernetesClientUtil;
 import org.apache.commons.io.FileUtils;
@@ -81,7 +80,6 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.eclipse.jkube.kit.common.util.KubernetesHelper.getCrdContext;
-import static org.eclipse.jkube.kit.common.util.KubernetesHelper.getFullyQualifiedApiGroupWithKind;
 import static org.eclipse.jkube.kit.common.util.KubernetesHelper.getKind;
 import static org.eclipse.jkube.kit.common.util.KubernetesHelper.getName;
 import static org.eclipse.jkube.kit.common.util.KubernetesHelper.getOrCreateLabels;
@@ -199,6 +197,10 @@ public class ApplyService {
             applyJob((Job) dto, sourceName);
         } else if (dto instanceof GenericCustomResource) {
             applyGenericCustomResource((GenericCustomResource) dto, sourceName);
+        } else if (dto instanceof Namespace) {
+            applyNamespace((Namespace) dto);
+        } else if (dto instanceof Project) {
+            applyProject((Project) dto);
         } else if (dto instanceof HasMetadata) {
             HasMetadata entity = (HasMetadata) dto;
             try {
@@ -1051,17 +1053,17 @@ public class ApplyService {
      */
     public boolean applyNamespace(Namespace entity) {
         String currentNamespace = getOrCreateMetadata(entity).getName();
-        log.info("Using currentNamespace: " + currentNamespace);
+        log.info("Creating currentNamespace: " + currentNamespace);
         String name = getName(entity);
         Objects.requireNonNull(name, "No name for " + entity );
         Namespace old = kubernetesClient.namespaces().withName(name).get();
         if (!isRunning(old)) {
             try {
                 Object answer = kubernetesClient.namespaces().create(entity);
-                logGeneratedEntity("Created currentNamespace: ", currentNamespace, entity, answer);
+                logGeneratedEntity("Created Namespace: ", currentNamespace, entity, answer);
                 return true;
             } catch (Exception e) {
-                onApplyError("Failed to create currentNamespace: " + name + " due " + e.getMessage(), e);
+                onApplyError("Failed to create Namespace: " + name + " due " + e.getMessage(), e);
             }
         }
         return false;
@@ -1087,7 +1089,7 @@ public class ApplyService {
             return false;
         }
         String currentNamespace = getOrCreateMetadata(entity).getName();
-        log.info("Using project: " + currentNamespace);
+        log.info("Creating project: " + currentNamespace);
         String name = getName(entity);
         Objects.requireNonNull(name, "No name for " + entity);
         OpenShiftClient openshiftClient = getOpenShiftClient();
@@ -1431,23 +1433,6 @@ public class ApplyService {
                 apply(entity, fileName);
             }
         }
-    }
-
-    /**
-     * Picks namespace for ApplyService, by default namespace specified in ClusterAccess is given
-     * preference. If there is namespace specified in provided Kubernetes resources, it would be
-     * given preference over ClusterAccess's namespace
-     *
-     * @param entities a list of Kubernetes resources to process
-     * @param clusterAccess {@link ClusterAccess} Kubernetes Cluster Access
-     * @return namespace for ApplyService
-     */
-    public static String getNamespaceForApplyService(Set<HasMetadata> entities, ClusterAccess clusterAccess) {
-        String namespaceInProvidedKubernetesList = KubernetesHelper.getNamespaceFromKubernetesList(entities);
-        if (namespaceInProvidedKubernetesList == null) {
-            return clusterAccess.getNamespace();
-        }
-        return namespaceInProvidedKubernetesList;
     }
 
     public static List<HasMetadata> getK8sListWithNamespaceFirst(Collection<HasMetadata> k8sList) {